Extract analyzer list
