+++ timesketch/frontend-ng/src/components/Analyzer/AnalyzerList.vue
@@ -1,0 +1,68 @@
+<!--
+Copyright 2021 Google Inc. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+
+<script>
+import TsSearchDropdown from '../Explore/SearchDropdown'
+import TsTimelineChip from '../Analyzer/TimelineChip'
+import ApiClient from '../../utils/RestApiClient'
+
+export default {
+  components:{
+    TsSearchDropdown,
+    TsTimelineChip,
+  },
+  data() {
+    return {
+      availableAnalyzers: [],
+    }
+  },
+  computed: {
+    sketch() {
+      return this.$store.state.sketch;
+    },
+  },
+  async created () {
+    this.availableAnalyzers = (await ApiClient.getAnalyzers(this.sketch.id)).data;
+  }
+
+}
+</script>
+
+<template>
+  <v-simple-table>
+    <template v-slot:default>
+      <thead>
+        <tr>
+          <th class="text-left">
+            Name
+          </th>
+          <th class="text-left">
+            Description
+          </th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr
+          v-for="analyzer in availableAnalyzers"
+          :key="analyzer.name"
+        >
+          <td>{{ analyzer.display_name}}</td>
+          <td>{{ analyzer.description }}</td>
+        </tr>
+      </tbody>
+    </template>
+  </v-simple-table>
+</template>
